+++ CHANGELOG.md
@@ -65,4 +65,8 @@ that the user is looking for);
 - Redesigned all "highlight" items;
 - Added "search history" button;
 
-- 
+- Added "search history" block;
+- Added search history system to weather app:
+    Users can see last 10 locations which they searched;
+    Users can search faster any location which they searched earlier;
+    Users can clear all search history;